requirements.txt
asgiref==3.5.2
Django==4.1.3
django-rest-auth==0.9.5
djangorestframework==3.14.0
Pillow==9.3.0
pygame==2.1.2
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
tzdata==2022.6
